adds WallGroup, aka wall vectors
Huge potential for performance improvement, due to the fact that
rendering walls is the most CPU-hungry task at the time being.

This update makes groups out of walls and blocks aligned on the X axis,
pre-renders them as a single tile, then remove all the tiles in the
group, except for the last one, which gets replaced with the
pre-rendered wall group.

There is a couple of issues remaining:

- Culling happens based on the position of the tile with the highest X
  coordinate, so walls tend to disappear even when they should still be
  on screen.

- Blocks and doorways. The whole wall gets rendered at once, meaning it
  is rendered OVER any door that would be located in one of the wall's
  doorways. This cause some visual issues.

1 is very hard to fix, 2 is more straightforward, but still requires
a lot of changes (simplest solution would be to prevent a wall group
from spanning over doorways).