Level data is no longer preemptively stored in game data files
This brings back an old feature from the 2010 project. A level's data is
only loaded when the level itself is first visited by the player. This
is useful for game editing: levels updated by the game editor will be
loaded in their latest version in older saves, as long as the player
hadn't been in these levels before. It will also soon be possible to add
a debugging function that resets a level to the initial state defined in
the main game data file (game.json).