implements QtGamepadLegacy
Due to changes related to the switch from X11 to Wayland, plus the fact
that Gamepad support was never successfully tested on Windows, and it's
very possible that Windows has similar restrictions as Wayland... due to
all of this, there are lots of chances involved with making gamepads
work again in Qt6.

This commit implements a VirtualCursor. But we like the system cursor,
which does not lag, regardless of the application's own charge, so we're
keeping BOTH the system cursor and a virtual cursor.
When the application itself moves the cursor, the system cursor is
hidden, and the virtual cursor appears in its place. When the system
cursor moves again, we do the reverse switcharoo.

We SHOULD use QCoreApplication::sendEvent to simulate mouse move events,
and while that works with most UIs, it does seem to break LevelMouseArea,
and I couldn't figure out why.
*So mouse move events are not simulated* yet !

Simulated clicks work great. That's a relief.

Don't forget to properly set QT_PLUGIN_PATH so the gamepad plugin loads.
I assume the default Windows plugin does not work, so you probably need to
build QtGamepadLegacy with the SDL2 plugin to make it work on Winwin.