update Level loading behavior
If unique characters are loaded after the level's TaskRunner is adjusted
to world time, then level effects that should be applied on a unique
character won't apply.

The characters can't be loaded before either, as
their own TaskRunner would be adjusted (and they don't need to, unique
characters are always in sync with the world time).

Hence this solution. If the level needs a time adjustment, then unique
characters are injected after the level's objects TaskRunners are
adjusted, and before the level's own TaskRunner is.